Student Sign Up Info Updated
Student Sign Up Info updated to match the requirements. Of the lab when you click the button in the signup page it just shows the information that the user entered. What will be added later is after putting in your info its supposed to take you to the exam page.